Split QrColorField and QrLogoInput out of qr.tsx into their own files
Reduced QrLogoInput's readFile (9->7 cyclomatic, 90->56 CRAP) by
extracting the SVG-vs-raster branching into prepareQrLogo(); also
parallelized its two independent awaits (rasterizing the image,
dynamic-importing the compression library) with Promise.all,
per react-doctor's server-sequential-independent-await check.

Extracted QrColorField's reset button and transparent-toggle into
ResetSwatchButton/TransparentToggle, and QrLogoInput's
busy/loaded/empty states into QrLogoButtonContent, plus a
RemoveLogoButton. Both components dropped out of fallow's flagged
list (7cyc/72CRAP and 7cyc/56CRAP -> 6cyc/42CRAP, no longer HIGH).

That extraction pushed qr.tsx to 7 components in one file, which
tripped react-doctor's no-multi-comp (Fast Refresh) check. Split
QrColorField and QrLogoInput (with their new sub-components and
logo-processing helpers) into qr-color-field.tsx and
qr-logo-input.tsx; qr.tsx now holds only QRPreview. readFile stays
HIGH (7cyc/56CRAP): the remaining branches are four independent
guard clauses (file/disabled/busy/org) plus a type check, which
belong together as one validation gate.

Verified: bun run check, worker tsc, bun test (74 pass), bun run
test:worker (46 pass), fallow (QrColorField/QrLogoInput both below
the HIGH threshold), react-doctor --scope changed (100/100, no
warnings, multi-comp resolved). Live-verified in the browser via a
throwaway pro-plan account: Settings' QR defaults card (color
swatches, transparent toggle + reset, logo dropzone) and the link
editor's QR customization panel both render and behave correctly;
test data cleaned up from D1 after.
